refactor(BRAZ-6): put the method GetByEmail in userRepository
