+++ services/users/mocks/user_repository_interface_mock.go
@@ -10,6 +10,9 @@
 package mocks
 
 import (
+	context "context"
+	reflect "reflect"
+
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -36,4 +39,18 @@ func NewMockUserRepositoryInterface(ctrl *gomock.Controller) *MockUserRepository
 // EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockUserRepositoryInterface) EXPECT() *MockUserRepositoryInterfaceMockRecorder {
 	return m.recorder
+}
+
+// GetByEmail mocks base method.
+func (m *MockUserRepositoryInterface) GetByEmail(ctx context.Context, collection, email string, structure any) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetByEmail", ctx, collection, email, structure)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// GetByEmail indicates an expected call of GetByEmail.
+func (mr *MockUserRepositoryInterfaceMockRecorder) GetByEmail(ctx, collection, email, structure any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetByEmail", reflect.TypeOf((*MockUserRepositoryInterface)(nil).GetByEmail), ctx, collection, email, structure)
 }
